Task #30

+++ App/controllers/competitor.py
@@ -2,7 +2,7 @@ from operator import attrgetter
 from App.models.competitor import Competitor
 from App.database import db
 import App.controllers.leaderboard as leaderboard
-# from App.controllers import create_message_Inbox
+import App.controllers.messageInbox as mi
 
 import App.controllers.messageInbox as mi
 
@@ -28,6 +28,10 @@ def get_competitor_by_username(username):
 def get_competitor(id):
     return Competitor.query.get(id)
 
+def get_competitor_json(id):
+    c =  Competitor.query.get(id)
+    return c.get_json()
+
 def get_all_competitors():
     return Competitor.query.all()
 
@@ -38,11 +42,12 @@ def get_all_competitors_json():
     competitors = [competitor.get_json() for competitor in competitors]
     return competitors
 
-def update_competitor(id, username, email):
+def update_competitor(id, username, email,password):
     competitor = get_competitor(id)
     if competitor:
         competitor.username = username
         competitor.email = email
+        competitor.password = password
         db.session.add(competitor)
         db.session.commit()
         return competitor
@@ -67,6 +72,10 @@ def  remove_competitor_overall_points(id,points):
         competitor = get_competitor(id)
         if competitor:
             competitor.overall_points = competitor.overall_points - points
+
+            if (competitor.overall_points < 0):
+                competitor.overall_points = 0
+
             db.session.add(competitor)
             db.session.commit()
             update_rank()
@@ -77,17 +86,22 @@ def  remove_competitor_overall_points(id,points):
         db.session.rollback()
 
 def delete_competitor(id):
+    
+    competitor = get_competitor(id)
     try:
-        competitor = get_competitor(id)
 
         if competitor:
-            db.session.delete(competitor)
-            db.session.commit()
-            leaderboard.populate_top20_leaderboards()
-            return True
+
+            message_inbox = mi.delete_message_inbox(id)
+            if message_inbox:
+                db.session.delete(competitor)
+                db.session.commit()
+                update_rank()
+                return True
         return False
     
     except Exception:
+        print("lo")
         db.session.rollback()
 
 def update_rank():
